Bitcamp25/retell_custom_functions/requirements.txt
requests
flask
flask_cors
retell-sdk